--- js/test.html
@@ -1,24 +1,0 @@
-<!DOCTYPE html>
-<html>
-  <head>
-    <meta charset="utf-8">
-  </head>
-  <body>
-    <button id="btn-1">
-  	<b>Нажмите</b>, чтобы узнать id кнопки!
-    </button>
-    <script>
-	  const getMessage = (elem) => {
-	    const id = elem.id;
-	    const message = `id = ${elem.id}`;
-	    return message;
-	  }
-	  const button = document.querySelector('#btn-1');
-	  button.addEventListener(`click`, (evt) => {
-	    const element = evt.target;
-	    const message = getMessage(element);
-	    alert(message);
-  	  });
-    </script>
-  </body>
-</html>
